refactor(plugin): extract runtime_ops.rs (#1079)
Moves the runtime-ops infrastructure (HOST_RUNTIME_TOKIO_HANDLE
static + install_host_runtime_tokio_handle + host_tokio_handle +
invoke_completion helper + CompletionGuard RAII type) plus the 7
host_rov_* callbacks + HOST_RUNTIME_OPS_VTABLE + accessor + the two
test modules out of host_services/mod.rs into a sibling submodule.
mod.rs re-exports HOST_RUNTIME_OPS_VTABLE, host_runtime_ops_vtable,
and install_host_runtime_tokio_handle (the latter is reached by
Runner::start at runtime startup).

cargo test -p streamlib-engine --lib unchanged at 1407 passed / 2
pre-existing Deno failures / 126 ignored.
